Creating an algorithm for playing like rock, paper scissors
We have to define a synchronize array of selection
For example array = ["paper", "rock", "scissors"]

Creating a function that compares two values based on specific rules:
If the first value is the first item in the array, and the second value is the last item, the second value wins.
If the first value is the last item in the array, and the second value is the first item, the first value wins.
Otherwise, the value with the smaller index in the array wins.